+++ Include/methodobject.h
@@ -39,7 +39,8 @@ PyAPI_FUNC(PyObject *) PyCFunction_Call(PyObject *, PyObject *, PyObject *);
 
 #ifndef Py_LIMITED_API
 PyAPI_FUNC(PyObject *) _PyCFunction_FastCallDict(PyObject *func,
-    PyObject **args, int nargs,
+    PyObject **args,
+    Py_ssize_t nargs,
     PyObject *kwargs);
 #endif
